requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow==15.0.0
dbt-duckdb==1.7.2
streamlit>=1.36.0
plotly